+++ processor/src/host/default.rs
@@ -1,4 +1,4 @@
-use alloc::{boxed::Box, sync::Arc, vec::Vec};
+use alloc::{sync::Arc, vec::Vec};
 
 use miden_core::{DebugOptions, Felt, Word, mast::MastForest};
 use miden_debug_types::{
@@ -74,29 +74,29 @@ where
         Ok(self)
     }
 
-    /// Loads a single [`EventHandler`] into this host.
+    /// Registers a single [`EventHandler`] into this host.
     ///
     /// The handler can be either a closure or a free function with signature
     /// `fn(&mut ProcessState) -> Result<(), EventHandler>`
-    pub fn load_handler(
+    pub fn register_handler(
         &mut self,
-        id: u32,
-        handler: impl EventHandler,
+        id: Felt,
+        handler: Arc<dyn EventHandler>,
     ) -> Result<(), ExecutionError> {
-        self.event_handlers.register(id, Box::new(handler))
+        self.event_handlers.register(id, handler)
     }
 
-    /// Unload a handler with the given id, returning a flag indicating whether a handler
+    /// Un-registers a handler with the given id, returning a flag indicating whether a handler
     /// was previously registered with this id.
-    pub fn unload_handler(&mut self, id: u32) -> bool {
+    pub fn unregister_handler(&mut self, id: Felt) -> bool {
         self.event_handlers.unregister(id)
     }
 
     /// Replaces a handler with the given id, returning a flag indicating whether a handler
     /// was previously registered with this id.
-    pub fn replace_handler(&mut self, id: u32, handler: impl EventHandler) -> bool {
+    pub fn replace_handler(&mut self, id: Felt, handler: Arc<dyn EventHandler>) -> bool {
         let existed = self.event_handlers.unregister(id);
-        self.load_handler(id, handler).unwrap();
+        self.register_handler(id, handler).unwrap();
         existed
     }
 
@@ -199,7 +199,7 @@ pub struct HostLibrary {
     /// A `MastForest` with procedures exposed by this library.
     pub mast_forest: Arc<MastForest>,
     /// List of handlers along with an event id to call them with `emit`.
-    pub handlers: Vec<(u32, Box<dyn EventHandler>)>,
+    pub handlers: Vec<(Felt, Arc<dyn EventHandler>)>,
 }
 
 impl From<Arc<MastForest>> for HostLibrary {