+++ processor/src/host/handlers.rs
@@ -1,6 +1,7 @@
 use alloc::{
     boxed::Box,
     collections::{BTreeMap, btree_map::Entry},
+    sync::Arc,
     vec::Vec,
 };
 use core::{error::Error, fmt, fmt::Debug};
@@ -33,6 +34,15 @@ where
 {
     fn on_event(&self, process: &ProcessState) -> Result<Vec<AdviceMutation>, EventError> {
         self(process)
+    }
+}
+
+/// A handler which ignores the process state and leaves the `AdviceProvider` unchanged.
+pub struct NoopEventHandler;
+
+impl EventHandler for NoopEventHandler {
+    fn on_event(&self, _process: &ProcessState) -> Result<Vec<AdviceMutation>, EventError> {
+        Ok(Vec::new())
     }
 }
 
@@ -91,7 +101,7 @@ pub type EventError = Box<dyn Error + Send + Sync + 'static>;
 /// ```
 #[derive(Default)]
 pub struct EventHandlerRegistry {
-    handlers: BTreeMap<u32, Box<dyn EventHandler>>,
+    handlers: BTreeMap<u64, Arc<dyn EventHandler>>,
 }
 
 impl EventHandlerRegistry {
@@ -99,13 +109,13 @@ impl EventHandlerRegistry {
         Self { handlers: BTreeMap::new() }
     }
 
-    /// Registers a boxed [`EventHandler`] with a given identifier.
+    /// Registers an [`EventHandler`] with a given identifier.
     pub fn register(
         &mut self,
-        id: u32,
-        handler: Box<dyn EventHandler>,
+        id: Felt,
+        handler: Arc<dyn EventHandler>,
     ) -> Result<(), ExecutionError> {
-        match self.handlers.entry(id) {
+        match self.handlers.entry(id.as_int()) {
             Entry::Vacant(e) => e.insert(handler),
             Entry::Occupied(_) => return Err(ExecutionError::DuplicateEventHandler { id }),
         };
@@ -114,8 +124,8 @@ impl EventHandlerRegistry {
 
     /// Unregisters a handler with the given identifier, returning a flag whether a handler with
     /// that identifier was previously registered.
-    pub fn unregister(&mut self, id: u32) -> bool {
-        self.handlers.remove(&id).is_some()
+    pub fn unregister(&mut self, id: Felt) -> bool {
+        self.handlers.remove(&id.as_int()).is_some()
     }
 
     /// Handles the event if the registry contains a handler with the same identifier.
@@ -129,9 +139,7 @@ impl EventHandlerRegistry {
         id: Felt,
         process: &ProcessState,
     ) -> Result<Option<Vec<AdviceMutation>>, EventError> {
-        if let Ok(id) = id.as_int().try_into()
-            && let Some(handler) = self.handlers.get(&id)
-        {
+        if let Some(handler) = self.handlers.get(&id.as_int()) {
             return handler.on_event(process).map(Some);
         }