+++ stdlib/src/lib.rs
@@ -1,13 +1,17 @@
 #![no_std]
 
+pub mod handlers;
+
 extern crate alloc;
 
 use alloc::{sync::Arc, vec, vec::Vec};
 
 use miden_assembly::{Library, mast::MastForest, utils::Deserializable};
 use miden_core::{Felt, Word};
-use miden_processor::HostLibrary;
+use miden_processor::{EventHandler, HostLibrary};
 use miden_utils_sync::LazyLock;
+
+use crate::handlers::keccak256::{KECCAK_HASH_MEMORY_EVENT_ID, handle_keccak_hash_memory};
 
 // STANDARD LIBRARY
 // ================================================================================================
@@ -32,7 +36,7 @@ impl From<&StdLibrary> for HostLibrary {
     fn from(stdlib: &StdLibrary) -> Self {
         Self {
             mast_forest: stdlib.mast_forest().clone(),
-            handlers: vec![],
+            handlers: stdlib.handlers(),
         }
     }
 }
@@ -45,6 +49,16 @@ impl StdLibrary {
     /// Returns a reference to the [MastForest] underlying the Miden standard library.
     pub fn mast_forest(&self) -> &Arc<MastForest> {
         self.0.mast_forest()
+    }
+
+    /// Returns a reference to the underlying [`Library`].
+    pub fn library(&self) -> &Library {
+        &self.0
+    }
+
+    /// List of all `EventHandlers` required to run all of the standard library.
+    pub fn handlers(&self) -> Vec<(Felt, Arc<dyn EventHandler>)> {
+        vec![(KECCAK_HASH_MEMORY_EVENT_ID, Arc::new(handle_keccak_hash_memory))]
     }
 }